+++ _Frontends/UI/main/src/service/LineManagementService.ts
@@ -44,6 +44,7 @@ export type StationData = {
     arrivingVehicle?: VehicleData,
     arrivingVehicleDistance?: number,
     arrivingVehicleStops?: number,
+    index?: number
 };
 export type VehicleData = {
     readonly entity: Entity,